+++ include/hs_routes.h
@@ -9,7 +9,9 @@
 #include "hs_routes_fs.h"
 #include "hs_routes_payload_limit.h"
 #include "hs_routes_powered_by.h"
+#include "hs_routes_redirection.h"
 #include "hs_routes_session.h"
+#include "hs_routes_static.h"
 
 #endif