+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### v0.4.1
 
-* Fix: Fix relative href values in FS directory route
+* Fix: relative href values in FS directory route
 
 ### v0.4.0 (2021-09-24)